merge master to arcade branch (#5424)
* Update to Onnxruntime 1.5.1 (#5406)

* Added variables to tests to control Gpu settings

* Added dependency to prerelease

* Updated to 1.5.1

* Remove prerelease feed

* Nit on GPU variables

* Change the _maxCalibrationExamples default on CalibratorUtils (#5415)

* Change the _maxCalibrationExamples default

* Improving comments

* Fix perf regression in ShuffleRows (#5417)

RowShufflingTransformer is using ChannelReader incorrectly. It needs to block waiting for items to read and was Thread.Sleeping in order to wait, but not spin the current core. This caused a major perf regression.

The fix is to block synchronously correctly - by calling AsTask() on the ValueTask that is returned from the ChannelReader and block on the Task.

Fix #5416

Co-authored-by: Antonio Velázquez <38739674+antoniovs1029@users.noreply.github.com>
Co-authored-by: Eric Erhardt <eric.erhardt@microsoft.com>